# Documentation/raw/swagger-transactions.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "EasyCard Transactions API",
    "version": "v1"
  },
  "paths": {
...
    "/api/cardtokens": {
      "get": {
        "tags": [
          "CardToken"
        ],
        "summary": "Get tokens by filters",
        "parameters": [
          {
            "name": "CreditCardTokenID",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "CardNumber",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "CardOwnerNationalID",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "CardOwnerName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "TerminalID",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "MerchantID",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "ConsumerID",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "ConsumerEmail",
            "in": "query",
            "description": "End-customer Email",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Take",
            "in": "query",
            "schema": {
              "minimum": -1,
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Skip",
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "SortBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortDesc",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "ShowDeleted",
            "in": "query",
            "description": "\r\n\r\nOnlyActive\r\n\r\nOnlyDeleted\r\n\r\nAll",
            "schema": {
              "$ref": "#/components/schemas/ShowDeletedEnum"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummariesResponse_CreditCardTokenSummary"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [
              "terminal_or_merchant_frontend_or_admin"
            ]
          }
        ]
      }
    },
    "/api/cardtokens/{key}": {
      "delete": {
        "tags": [
          "CardToken"
        ],
        "summary": "Delete credit card token",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationResponse"
                },
                "example": {"message":"Validation Errors","status":"error","correlationId":"1898bf3a-70b3-4a9b-b7b2-432af48a53de","errors":[{"code":"paymentRequestAmount","description":"Could not convert string to decimal: 1200$. Path 'paymentRequestAmount', line 10, position 33."}]}
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [
              "terminal_or_merchant_frontend_or_admin"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CardExpiration": {
        "required": [
          "month",
          "year"
        ],
        "type": "object",
        "properties": {
          "year": {
            "maximum": 90,
            "minimum": 20,
            "type": "integer",
            "format": "int32"
          },
          "month": {
            "maximum": 12,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "expired": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "CreditCardTokenSummary": {
        "type": "object",
        "properties": {
          "creditCardTokenID": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "terminalID": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "merchantID": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "cardNumber": {
            "type": "string",
            "nullable": true
          },
          "cardExpiration": {
            "$ref": "#/components/schemas/CardExpiration"
          },
          "cardVendor": {
            "type": "string",
            "nullable": true
          },
          "created": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "cardOwnerName": {
            "type": "string",
            "nullable": true
          },
          "consumerID": {
            "type": "string",
            "description": "Consumer ID",
            "format": "uuid",
            "nullable": true
          },
          "consumerEmail": {
            "type": "string",
            "description": "End-customer Email",
            "nullable": true
          },
          "expired": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "OperationResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/StatusEnum"
          },
          "entityID": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "entityUID": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "entityReference": {
            "type": "string",
            "nullable": true
          },
          "correlationId": {
            "type": "string",
            "nullable": true
          },
          "entityType": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Error"
            },
            "nullable": true
          },
          "concurrencyToken": {
            "type": "string",
            "nullable": true
          },
          "innerResponse": {
            "$ref": "#/components/schemas/OperationResponse"
          },
          "additionalData": {
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ShowDeletedEnum": {
        "enum": [
          "OnlyActive",
          "OnlyDeleted",
          "All"
        ],
        "type": "string",
        "description": "\r\n\r\nOnlyActive\r\n\r\nOnlyDeleted\r\n\r\nAll",
        "x-enumNames": [
          "OnlyActive",
          "OnlyDeleted",
          "All"
        ]
      },
...
      "StatusEnum": {
        "enum": [
          "success",
          "warning",
          "error"
        ],
        "type": "string",
        "description": "\r\n\r\nsuccess\r\n\r\nwarning\r\n\r\nerror",
        "x-enumNames": [
          "Success",
          "Warning",
          "Error"
        ]
      },
...
      "SummariesResponse_CreditCardTokenSummary": {
        "type": "object",
        "properties": {
          "numberOfRecords": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreditCardTokenSummary"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "oauth2": [ ]
    }
  ]
}